# FastRegex: Skip simplify step if it makes the regexp a lot longer 