I noticed that in some cases the preparation to optimise a regexp in `FastRegexMatcher` can get costly.
This adds a small check to see if the `Simplify` step has made the regexp a lot longer, in which case we will scan the original string.

#### Does this PR introduce a user-facing change?
```release-notes
NONE
```

```
goos: linux
goarch: amd64
pkg: github.com/prometheus/prometheus/model/labels
cpu: Intel(R) Core(TM) i7-14700K
                                         │    before    │               after                │
                                         │    sec/op    │   sec/op     vs base               │
NewMatcher/regex_matcher_with_literal-28    78.73n ± 0%   78.71n ± 1%        ~ (p=0.937 n=6)
NewMatcher/complicated_regex-28            1226.3µ ± 2%   457.0µ ± 4%  -62.73% (p=0.002 n=6)
FastRegexMatcher/(a||b){1000}-28            480.1µ ± 2%   471.2µ ± 1%   -1.85% (p=0.009 n=6)
geomean                                     35.92µ        25.69µ       -28.49%

                                         │    before    │                  after                  │
                                         │     B/op     │      B/op       vs base                 │
NewMatcher/regex_matcher_with_literal-28     240.0 ± 0%     240.0 ±   0%        ~ (p=1.000 n=6) ¹
NewMatcher/complicated_regex-28            2.965Mi ± 0%   1.833Mi ±   0%  -38.19% (p=0.002 n=6)
FastRegexMatcher/(a||b){1000}-28           14706.5 ± 9%     325.5 ± 100%  -97.79% (p=0.002 n=6)
geomean                                    21.70Ki        5.190Ki         -76.08%
¹ all samples are equal

                                         │   before    │                after                │
                                         │  allocs/op  │ allocs/op   vs base                 │
NewMatcher/regex_matcher_with_literal-28    5.000 ± 0%   5.000 ± 0%        ~ (p=1.000 n=6) ¹
NewMatcher/complicated_regex-28            9170.0 ± 0%   122.0 ± 0%  -98.67% (p=0.002 n=6)
FastRegexMatcher/(a||b){1000}-28            1.000 ± 0%   1.000 ±  ?        ~ (p=0.455 n=6)
geomean                                     35.79        8.481       -76.30%
¹ all samples are equal
```